requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
crispy-tailwind==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.5
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.11.0
django-environ==0.4.5
django-phonenumber-field==5.0.0
Faker==5.6.5
gunicorn==20.0.4
idna==2.10
migrate==0.3.7
misaka==2.1.1
oauthlib==3.1.0
phonenumbers==8.12.18
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
whitenoise==5.2.0
